Feat: 개인정보동의 팝업창 컴포넌트화(어휴 깨끗해..)
